+++ BMS/Note.cs
@@ -4,12 +4,12 @@
     public class Note : ICloneable
     {
         public double Section { get; private set; }
-        public long Time { get; private set; }
+        public double Time { get; private set; }
         public int Wav { get; private set; }
-        public long Start { get; private set; }
-        public long Duration { get; private set; }
+        public double Start { get; private set; }
+        public double Duration { get; private set; }
         public int State { get; private set; }
-        public long PlayTime { get; private set; }
+        public double PlayTime { get; private set; }
         private Note[] layeredNotes = [];
         public Note[] LayeredNotes {
             get 
@@ -21,10 +21,10 @@
         public void SetSection(double section) => Section = section;
         public void SetWav(int wav) => Wav = wav;
         public void SetState(int state) => State = state;
-        public void SetStartTime(long time) => Start = time;
-        public void SetDuration(long duration) => Duration = duration;
+        public void SetStartTime(double time) => Start = time;
+        public void SetDuration(double duration) => Duration = duration;
         public void SetPlayTime(int playtime) => PlayTime = playtime;
-        public void SetTime(long time) => Time = time;
+        public void SetTime(double time) => Time = time;
 
         public void AddLayeredNote(Note n)
         {
